Padrão de projeto Template Method e Interface Segregation Principle (SOLID).
